Added plate image with initial attempt to display cup alongside, restructured image paths, renamed and moved some files
