+++ app/App.js
@@ -1,13 +1,12 @@
 import React, { useState } from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-
-import RootNavigation from './navigation/root-navigator'
+import RootNavigation from './navigation/RootNavigator'
 
 export default function App() {
   const [appTitleName] = useState("Meal Planner");
   return (
     <NavigationContainer>
-        <RootNavigation />
+      <RootNavigation />
     </NavigationContainer>
   ); 
 }
