+++ app/components/drinks.js
@@ -1,10 +1,16 @@
 import * as React from 'react';
-import { Text, View, StyleSheet } from 'react-native';
+import { Text, View, StyleSheet, ImageBackground } from 'react-native';
 
 const Drink = (props) => {
   return (
     <View style={styles.container}>
       <Text>{props.name}</Text>
+      <ImageBackground
+        source={require(`../assets/images/cup/Cup-empty.png`)}
+        // source={require(props.imagePath)}
+        style={{width: 150, height: 150}}
+      >
+      </ImageBackground>
     </View>
   );
 }
