+++ app/components/screens.js
@@ -1,16 +1,6 @@
 import * as React from 'react';
 import { Text, View } from 'react-native';
-import { Mug, Glass } from '../components/drinks'
-
-const Plate = (props) => {
-  return (
-    <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>{props.name} Plate!</Text>
-      <Mug />
-      <Glass />
-    </View>
-  );
-}
+import { Plate } from './Plate'
 
 function BreakfastPlate() {
   return (
